+++ lib/autohttp/agent.js
@@ -96,22 +96,20 @@ class AutoHttp2Agent extends EventEmitter {
       return
     }
 
-    if (protocol === 'http/1.1' || protocol === 'http/1.0') {
-      const http1RequestOptions = Object.assign({}, options, {
-        agent: this.httpsAgent
-      })
+    const http1RequestOptions = Object.assign({}, options, {
+      agent: this.httpsAgent
+    })
 
-      let request
-      try {
-        request = https.request(http1RequestOptions)
-      } catch (e) {
-        cb(e)
-        return
-      }
-
-      request.on('socket', (socket) => socketCb(socket))
-      cb(null, 'http1', request)
+    let request
+    try {
+      request = https.request(http1RequestOptions)
+    } catch (e) {
+      cb(e)
+      return
     }
+
+    request.on('socket', (socket) => socketCb(socket))
+    cb(null, 'http1', request)
   }
 
   createNewSocketConnection (req, options, cb, socketCb) {
