+++ lib/autohttp/headerValidations.js
@@ -22,7 +22,19 @@ function checkIsHttpToken (token) {
   return RegExp(tokenRegExp).exec(token) !== null
 }
 
+// Referenced from https://github.com/nodejs/node/blob/0bf200b49a9a6eacdea6d5e5939cc2466506d532/lib/internal/http2/core.js#L1763
+function validateRequestHeaders(headers) {
+  if (headers !== null && headers !== undefined) {
+    const keys = Object.keys(headers)
+    for (let i = 0; i < keys.length; i++) {
+      const header = keys[i]
+      if (header[0] === ':') {
+        assertValidPseudoHeader(header)
+      } else if (header && !checkIsHttpToken(header)) { throw new Error('Invalid HTTP Token: Header name' + header) }
+    }
+  }
+}
+
 module.exports = {
-  assertValidPseudoHeader,
-  checkIsHttpToken
+  validateRequestHeaders
 }
