+++ lib/autohttp/request.js
@@ -1,7 +1,7 @@
 const { EventEmitter } = require('events')
 const { Http2Request: HTTP2Request } = require('../../lib/http2/request')
 const { globalAgent } = require('./agent')
-const { assertValidPseudoHeader, checkIsHttpToken } = require('./headerValidations')
+const { validateRequestHeaders } = require('./headerValidations')
 
 const kJobs = Symbol('kJobs')
 
@@ -9,8 +9,6 @@ class MultiProtocolRequest extends EventEmitter {
   constructor (options) {
     super()
     this[kJobs] = []
-    this.onHttp2 = this.onHttp2.bind(this)
-    this.onHttp = this.onHttp.bind(this)
     this.options = options
     this.options.host = options.hostname || options.host || 'localhost'
 
@@ -182,17 +180,7 @@ function request (options) {
   // HTTP/2 internal implementation sucks. In case of an invalid HTTP/2 header, it destroys the entire session and
   // emits an error asynchronously, instead of throwing it synchronously. Hence, it makes more sense to perform all
   // validations before sending the request.
-  const headers = options.headers
-
-  if (headers !== null && headers !== undefined) {
-    const keys = Object.keys(headers)
-    for (let i = 0; i < keys.length; i++) {
-      const header = keys[i]
-      if (header[0] === ':') {
-        assertValidPseudoHeader(header)
-      } else if (header && !checkIsHttpToken(header)) { throw new Error('Invalid HTTP Token: Header name ' + header) }
-    }
-  }
+  validateRequestHeaders(options.headers)
 
   return new MultiProtocolRequest(options)
 }
